Attempt to reconnect the database
The connection to the Postgres database seems to drop out sometimes, possibly due to firewall settings between the app and the database, or TCP keepalive settings, or some other unknown factor. To work around that problem, check if the database connection has been closed, and if so, try to reconnect to it.